Add files via upload
I tried to use a different heuristic. In particular, after computing the number of uncovered elements for a given state, for each of the unused sets I compute the number of uncovered elements they are able to cover, separately for each set (improvement level). Then count how many sets are necessary such that the sum of improvement level is equal or greater than the number of uncovered elements.